feat: Create first Proof of Concept
It consist of two projects, a console and a web server.

The web server deliver a simple html page on http://localhost:5000

The console application ask the user a URI and a word.
Then it count how many time the word appear in the source code.
It does not render a web page, it only GET the URI